Keep pinned Node runtime management together
Amp-Thread-ID: https://ampcode.com/threads/T-01a060b5-afe6-7778-bcf7-a5e7c79bd3dd
